fix(go-processor): change MediaObject.Duration to float64
Telegram sends fractional durations (e.g. 16.75s). The *int type caused
json.Unmarshal to fail silently, skipping the entire image_url extraction.

+++ services/go-processor/internal/domain/models.go
@@ -113,5 +113,5 @@ type MediaObject struct {
 	PreviewURL *string `json:"preview_url,omitempty"`
 	Width     *int    `json:"width,omitempty"`
 	Height    *int    `json:"height,omitempty"`
-	Duration  *int    `json:"duration,omitempty"`
+	Duration  *float64 `json:"duration,omitempty"`
 }
